Only rexecute do_tag if filter is not empty
Problem:
Enter a search query that does not return any threads. Then trigger
search_tag, e.g. by pressing "a" to archive the current thread.
thread_id is now not available, the filter is empty. In tha case
db.query() will match all messages in the database. As a result all
massages get tagged with the passed tags (in case of 'a' all messages
will get archived - tagged as "-unread -inbox")

Solution:
Only run the body of do_tag if filter is not empty.

+++ plugin/notmuch.rb
@@ -243,24 +243,26 @@ def search_render(search)
 end
 
 def do_tag(filter, tags)
-  $curbuf.do_write do |db|
-    q = db.query(filter)
-    q.search_messages.each do |e|
-      e.freeze
-      tags.split.each do |t|
-        case t
-        when /^-(.*)/
-          e.remove_tag($1)
-        when /^\+(.*)/
-          e.add_tag($1)
-        when /^([^\+^-].*)/
-          e.add_tag($1)
+  if not filter.empty?
+    $curbuf.do_write do |db|
+      q = db.query(filter)
+      q.search_messages.each do |e|
+        e.freeze
+        tags.split.each do |t|
+          case t
+          when /^-(.*)/
+            e.remove_tag($1)
+          when /^\+(.*)/
+            e.add_tag($1)
+          when /^([^\+^-].*)/
+            e.add_tag($1)
+          end
         end
+        e.thaw
+        e.tags_to_maildir_flags
       end
-      e.thaw
-      e.tags_to_maildir_flags
+      q.destroy!
     end
-    q.destroy!
   end
 end
